back/pyproject.toml
[tool.poetry]
name = "capstone-back"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aniso8601 = "9.0.1"
asgiref = "3.6.0"
django = "4.1.5"
djangorestframework = "3.14.0"
graphene = "3.2.1"
graphene-django = "3.0.0"
graphql-core = "3.2.3"
graphql-relay = "3.2.0"
mysqlclient = "2.1.1"
promise = "2.3"
pytz = "2022.7.1"
six = "1.16.0"
sqlparse = "0.4.3"
text-unidecode = "1.3"
tzdata = "2022.7"
strawberry-graphql = "0.159.0"
django-cors-headers = "^3.13.0"
strawberry-graphql-django = "^0.9.2"
strawberry-django-plus = "^2.0.4"
python-dotenv = "^0.21.1"
fastapi = "^0.92.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
gunicorn = "^20.1.0"
setuptools = "^67.4.0"
autopep8 = "^2.0.2"
django-debug-toolbar = "^3.8.1"
sentry-sdk = "^1.18.0"
django-choices-field = "^2.1"
boto3 = "^1.26.127"

[tool.autopep8]
max_line_length = 120
ignore = "E501,E501,F403,F401,F821"
in-place = true
recursive = true
aggressive = 3


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
